root/app/requirements.txt
uvicorn
fastapi
aiofiles
Jinja2
python-multipart
importlib_metadata
nltk
tqdm
pyautogui
pillow